requirements.txt
streamlit
openai
requests
bs4
anthropic
google-generativeai

PyPDF2
pyngrok
fitz
frontend
ngrok
pymupdf
tiktoken